Render split refactor
